+++ steam/client/cdn.py
@@ -5,12 +5,12 @@ Initializing :class:`.CDNClient` requires a logged in :class:`.SteamClient` inst
 
 .. warning::
     This module uses :mod:`requests` library, which is not gevent cooperative by default.
-    It is high recommended that you use :meth:`steam.client.monkey.patch_minimal()`.
+    It is high recommended that you use :meth:`steam.monkey.patch_minimal()`.
     See example below
 
 .. code:: python
-    import steam.client.monkey
-    steam.client.monkey.patch_minimal()
+    import steam.monkey
+    steam.monkey.patch_minimal()
 
     from steam.client import SteamClient, EMsg
     from steam.client.cdn import CDNClient